Fix : BoardContentNo의 타입이 잘못되어있던 점 수정

+++ back/kinder/src/main/java/com/bridge/kinder/dto/BoardContentDto.java
@@ -62,7 +62,7 @@ public class BoardContentDto {
     @AllArgsConstructor
     @Builder
     public static class Update {
-        private Integer contentId;  // 기존 content 구분용 ID
+        private Long contentId;  // 기존 content 구분용 ID
         private CommonEnums.BoardContentType type;
         private String contentText;
         private String contentFile; // 수정된 경우 경로 저장
